Bug #18842228 SERVER NOT ACCEPTING SOME CONNECTIONS WHEN OFFLINE_MODE IS ON
The server runs into a race condition and occasionally rejects
SUPER connections when OFFLINE_MODE is TRUE.

Analysis:
When OFFLINE MODE is set to TRUE the server goes through the list of
threads and kills all the threads which are not SUPER. A new thread is
first added to the list of threads and then after short while it gets
its permissions assigned. Thus there can be super threads on the thread
list without super permission assigned yet and therefore those threads
will be killed.
Fix:
Kill only the threads that have account (permissions) assigned. If the
thread has not yet account assigned and it turns out that the thread is
not super the authentication for this thread will fail as OFFLINE_MODE
is true.