Bug#20422145 DO ABI CHECK FOR SERVICES SEPARATELY FROM PLUGINS
The file include/mysql/plugin.h currently does #include <mysql/services.h>
This means that the ABI check for *all* plugins,
depend on *all* services.

If you want to add a new service, lots .h.pp files will need to be rewritten.
This doesn't scale and is hard to maintain if we want to add more plugin
services.

Fix:
Remove that #include when doing ABI check
Keep it when *not* doing ABI check.
This allows plugin implementors to simply do
  include <mysql/plugin.h>
to get access to all plugin services.

Write a separate ABI check for services.h