BUG#20445758 - GROUP REPLICATION PLUGIN SHOULD NOT PASS INTERNAL POINTERS TO P_S TABLES
Description
============

Group Replication stats and status are available to users trough
performance schema tables.

Those performance schema tables are defined on server, and when user
selects them the tables will ask the needed data to plugin using the
plugin interface.

The problem is that some of the returned info from plugin are internal
pointers that caller can modify, to avoid that all data must be returned
as a copy instead of internal pointers.

Solution
============

All returned and default values to P_S tables now belong
to the Group Replication plugin, which returns them via output pointer
values, making a snapshot of the current values.