Bug #18845301    THERE IS A RACE CONDITION IN BINLOG GROUP COMMIT AND PERFORMANCE SCHEMA INSTRUMENT
Problem Description:

Lets imagine following case:
We have two threads which are about to commit. First one becomes a
leader and
    if (change_stage(thd, Stage_manager::COMMIT_STAGE,
                   final_queue, &LOCK_sync, &LOCK_commit))

returns false for it and it proceeds to process_commit_stage_queue.

Second one is follower and it goes to
    if (!leader)

branch in Stage_manager::enroll_for to wait until commit done by the
leader.

What leader does in process_commit_stage_queue is trying to attach
to the follower (replace leader's current_thd and PSI_thread with
follower's) and execute ha_commit_low.

There is a race condition here. Follower executes PFS
instrumentation code when acquires locks and waits on condition
variables (in !leader branch). ha_commit_low also uses
mutexes/rwlocks. The do it concurrently with the same PSI_thread
thread-specific variable set.

Concurrent uses of the same PFS instrumentation for *_cond_wait and *_mutex_wait
is not supported and is not thread-safe. For example there are unprotected statements
thread->m_events_waits_current++/thread->m_events_waits_current--
which can cause thread->m_events_waits_current to went out of bounds
of thread->m_events_waits_stack and server crash.


Fix: In the old code, the commit actions performed by the leader
(on behalf of follower) are indicated in PFS instrumentation in
the name of the follower which is *not* indicating the real statistics
happening inside the system. Now the code is changed to make sure that
the PFS instrumentation will exactly give the information what is happening
inside the system i.e., Leader work (all the commit work) will be accounted
on leader's name and followers work (mostly waiting on mutex)  be accounted
on follower's name. This can be achieved by simply ignoring PSI context
change when the leader is trying to attach to follower's context i.e., leader
will reach PFS with it's own PSI_thread object instead of follower's
PSI_thread object when it is doing commit on behalf of the follwer.