Bug#20460208: !table || (!table->read_set || bitmap_is_set)
The optimizer relies on thd->mark_used_columns to have the value
MARK_COLUMNS_READ for columns in WHERE conditions, etc.
If this is not the case, fields are sometimes erroneously marked for
write or not marked at all. In particular, if a failed statement of
some "wrong" type preceded a correct statement, this problem might hit.
In the bug report, a failed INSERT statement is followed by a DELETE
statement.

This is a quick fix that should provide safeguards for DELETE and
UPDATE statements. It sets mark_used_columns to MARK_COLUMNS_READ
when needed and restores the old value after use.

A more thorough fix with complete safeguard will follow later.