BUG#20399899 BINLOG CORRUPTION. CLIENT: ERRNO 3. SLAVE: ERRNO              1236 AND 1594
There was a logical error in MYSQL_BIN_LOG::do_write_cache().

if (buf_len < LOG_EVENT_HEADER_LEN)
{
  ...
  handle the case that event header is splitted into two buffers.
}

The condition forgot to check if event_len is 0, so for below situation:
- The transaction's binlog events are larger than binlog_cache_size.
- The last event is splitted into two buffers and the part in second buffer
  is smaller than LOG_EVENT_HEADER_LEN.

It is not a event header but it is handled as a event header wrongly.

This patch fixed it, and also simplified the code in do_write_cache().
Change to binlog.cc credited to Victor Kirkebo.