BUG#20455917: GROUP REPLICATION APPLIER CHANNEL IS NOT LISTED ON ALL REPLICATION P_S TABLES
Group Replication applier channel was not being listed on all
replication performance schema tables, what was causing some status
values not to be accessible.
This was happening because Group Replication applier channel doesn't
have a regular Master_info object, since it doesn't have a IO
thread. Applier data comes from group communication toolkit and not
from a point-to-point connection from a master.

On this patch we ensure that Group Replication applier channel is
listed on all replication performance schema tables by adding the
channel with a fake master hostname "<NULL>".