BUG#20007583: THE EVENT_SCHEDULER USERNAME IS NOT RESERVERD.               ALLOWS PROCESSLIST VIEW
Analysis
--------
A user created with the name 'event_scheduler' can
view the process list of the event scheduler without
having the PROCESS privilege.

When a SHOW PROCESSLIST query is issued, if the user
does not have the PROCESS privileges, only the threads/
connections spawned by the user is displayed. This is
determined by comparing the user name of all threads
with the user name who issued the SHOW PROCESSLIST
query. Since the 'event_scheduler' system thread sets
the user name as 'event_scheduler', the comparison
succeeds with user 'event_scheduler', thus displaying
the event_scheduler system thread entry as well.

Fix:
---
Avoid showing information of the system threads if the
user does not have the PROCESS privilege.