BUG#19982543 SERIOUS TPS DECLINE IF GTID IS ENABLED IN 5.7
The GTIDs of transactions committed in group weren't added to
GTID_EXECUTED in order. Because of this, sometimes it created
temporary gaps in the GTID_EXECUTED which lead to more contention
than necessary.

When this gaps happen in GTID_EXECUTED, the server would have to add
and remove intervals from the Gtid_set, and adding and removing
intervals requires a mutex, which would cause contention.

Fix: moved the code that handled the GTIDs to a place where it will
be added to GTID_EXECUTED in the same commit order, without gaps.