Bug#20381342 EXPIRATION TIME IGNORED
This bug is caused by the expired time set to a interval value, not
a time value.
For example, if the expired time is set to 60 secs, which means after 60
secs from current time, the item will be expired.
So, the expired time of item should be current time + 60 secs, not 60
secs.

Reviewed-by: Jimmy Yang<jimmy.yang@oracle.com> on IM.