Bug #19323016 FIX FOR BUG#19056196 YIELDS WRONG RESULT.
Calling the zero-argument version of the function is defined to give
the same result as calling the single-argument version with the
context item (.). That is, fn:number() is equivalent to fn:number(.).

This means that these two queries

  SELECT ExtractValue('<a><b/></a>', 'number(.)');

and

  SELECT ExtractValue('<a><b/></a>', 'number()');

should produce the same result (0)

Analysis: number function should use agument if availabe. If it is not
available then it should use context if available and if it the context is
not available too it should use the root node.