Bug#20471685 MYSQLD CRASHES DURING WRITE TESTING FOR ABOVE 1000 CONNECTIONS
Problem:

The code implemented for
  WL#7794 PERFORMANCE_SCHEMA SCALABLE MEMORY ALLOCATION,
in
  PFS_buffer_scalable_container::allocate(),
is not thread safe.

The original intent was to have:
- thread 1 detects a buffer needs to be enlarged
- thread 2 detects a buffer needs to be enlarged
- thread 1 creates a new page and increase the number of pages
- thread 2 reuse the page just created by thread 1.

However in the code, there is a window of operation when:
- the number of pages in already incremented
- the new page is not yet added to the m_pages[] array.
which creates a crash, by using a NULL page in thread 2

Solution:

The fix is to revise PFS_buffer_scalable_container::allocate()
to rely on a single atomic compare and swap, to:
- either add a new page
- or reuse a new page added by a concurent thread
so that this works properly concurrently.

Only after a page is added, increment the number of pages.