Bug #19723750 MYSQL --QUICK CAUSES SEGMENTATION FAULT
Description:
It seems any command entered interactively in the mysql client with the
--quick option causes the clause to crash with a segmentation fault.

Analysis:
As per documentation - with quick option no history gets written to the
file. The history was not initialized for quick, however the history
writes were independed of quick value.
Fix:
Made history writes depend on quick variable.