BUG#20478089: SUPPRESS VIEW_CHANGE_LOG_EVENT MISSING MTS SEQUENCE NUMBER
When logical clock slave parallel applier is ON, binary log events
fields sequence_number and last_committed are used to control the window
on which events may be applied in parallel.
These fields are set on Gtid_log_event for all transactions/statements.
View_change_log_event, which marks a member join or leave from Group
Replication membership, is a special event that doesn't have
sequence_number and last_committed fields and should be treated as a
single transaction. So the warning thrown when a View_change_log_event
is read from relay log is misleading and should be suppressed for
that type.

Removed the warnings of missing sequence_number and last_committed
on View_change_log_event.
Also improved channel_get_appliers_thread_id() function pointers
handling.