Bug #20267451	- WL#8242: ERROR - UNABLE TO BROADCAST STABLE TRANSACTIONS SET MESSAGE
Problem
===============
In a group replication setup , in one of the servers the following error is
occurring (in the error log):

[..]
2014-12-24T08:30:07.914076Z 0 [ERROR] Plugin gcs_replication_plugin reported:
'Unable to broadcast stable transactions set message'
[..]

The running test hangs and the server halts.


Analysis
===============
After using a simpler version of the test provided by Namit, one was
able to trace the problem to parallel transactions on the same node.
Digging deep into the code, the node hangs when both transactions call
send_message at the same time. This enters the Corosync realm of IPC
between the client and a server which seem not to be thread-safe when
waiting for an asynch answer.

Solution
===============
Synchronize the calls to the public send_message method of the Binding
interface implementation of Corosync. It is the smaller possible grain
to avoid interference in the algorithm.