Bug#19856162: HANDLE_FATAL_SIGNAL 11 IN PARTITION_INFO::CHECK_PARTITION_INFO
Problem in prep_alter_part_table after error,
result in altered partition_info object was still used.

Solution was to only work on a clone and only use that
clone when needed and switch back to the original
partition_info object afterwards, so the altered object
would not be used in other cases (or be reused by
following statements under LOCK TABLES).