Add missing failure detection for 'testBasic -n Bug54986'.
The test continued on after some operation failed,
which finaly caused it to report 'match failure' as the
DB did not contain what the test framework expected to find.

Still expect test to fail after this fix, but now it will
hopefulle be possible to pinpont the point in time when it failed.