Bug20448413: MSR+MTS+SLAVE_TRANSACTION_RETRIES TRIES TO              OPEN RELAY LOG WHICH DOESN'T EXIST
Bug has arised because of improper merging of wl#6964 to multisource code.
During a transaction retry, a slave worker constructs the relay log path
of the event by combining the relay_log_base_name with the number extension.
With multisource replication is on, the worker incorrectly constructs relay a
log file path for it's channel.
The present code correctly fixes the relay log file path by a channel worker
by adding the channel name to the relay_log_base_name