Bug#19828320 FTS MATCH IN SUBQUERY CRASH AFTER UPDATE_REF_AND_KEYS
The fulltext predicate is inside a subquery and involves
an outer reference; it thus cannot be used for FT index lookup,
but MySQL does not see it, which causes a crash.
Fix: imitate the existing solution used for ordinary index lookups,
where we don't create a Key_use if the field is outer
(see call to is_local_field() in add_key_fields()).
This patch is based on work by Sergey Glukhov.