Bug#20210742 GROUP BY ON MERGE VIEW WITH ORDER BY DOES NOT WORK WITH ONLY_FULL_GROUP_BY
Before WL#5275, the view's ORDER BY was merged into the outer query,
giving SELECT d,COUNT(*) FROM t1 ... GROUP BY d ORDER BY n;
which was rejected because 'n' is not functionally dependent on 'd'.
With WL#5275, ORDER BY is not merged anymore (as the outer query
is grouped), so the problem is solved. Just adding the testcase.