Bug#20030284 - HANDLE_FATAL_SIGNAL (SIG=11) IN NAME_HASH_SEARCH
Issue here is, "table_name" passed to function name_hash_search()
is "NULL". The "table_name"(NULL) is used as a source argument
in strcpy operation. This is resulting in the issue reported in
the bug report.

While processing derived table needed for the query execution,
the name of the table (i.e. TABLE_LIST::table_name) is made to
point to the temporary table name held in "TABLE_SHARE::
table_name.str". Member "TABLE_LIST::table_name" with temporary
table name is used through out the execution of the query. While
closing "THD::derived_tables", the instances of TABLE and TABLE_SHARE
allocated are deleted. Because of this "TABLE_LIST::table_name"
becomes stale pointer. While initializing prepared statement for the
next execution, "TABLE_LIST::table_name" and "TABLE_LIST::db" is reset
to NULL for the derived tables. And while opening the tables of prepared
statement, "TABLE_LIST::table_name" and "TABLE_LIST::db" of derived
table created for view is restored from the "TABLE_LIST::view_name"
and "TABLE_LIST::view_db".

But for the prepared statement, create_view_precheck() is called
before restoring the "table_name" and "db" of derived table
created for the view. Because of this NULL value is passed to the
"name_hash_search" function.

Fix:
---------------
To fix this issue, added a code to restore table name and database
name of the derived tables created for view during the initialization
of prepared statement.
Logic to restore the table name and database name in open table logic
is required for the stored routines. So added a check in open
tables logic to not restore table name and database name again if
it is already restored.