Bug#20416705 ASSERTION `M_SRID_FOUND_IN_DOCUMENT == -1' FAILED
ANALYSIS
------------
When evaluating a field with ST_GeomFromGeoJSON(), the function
expects m_srid_found_in_document to be set to the default value
of -1 before starting the evaluation. The function never resets
this value to the default, and an assertion will thus fail when
this function is for selecting more than one GeoJSON field with
valid CRS data.

SOLUTION
--------
Reset the variable m_srid_found_in_document to -1 before
evaluating the GeoJSON input.