Bug#20454533: Assertion failed: sargables == 0 || *keyfields ...
The fix for "regression from bug#19789450" had a problem for some
non-SELECT statements: These statements call st_select_lex::prepare()
after preparing some items specific to the UPDATE/DELETE operation,
so the prepare() call is clearing cond_count set for such references.
When a derived table is involved in the UPDATE, conditions counted in
the derived table are propagated to the outer query block. This happens
before cond_count is cleared in ::prepare().

The fix is to move the resetting of cond_count, etc from ::prepare()
to reinit_stmt_before_use(). Thus, the fields are reset by the
st_select_lex constructor for the first ::prepare(), for the
subsequent ::prepare()s, they are reset by reinit_stmt_before_use()