Description ===============
Using a test that stresses Group Replication startup (START and STOP
sequence), one can see a startup crash.

Analysis
===============

In a typical Group Replication Test, one uses start_group_replication.inc
that starts the plugin and waits for Recovery to finish to declare the
node online.

This test case stresses this critical step by starting and stopping
without waiting. As such, it causes unexpected side effects in the
Group Replication binding such as unexisting views and antagonic
State Exchange processes running at the same time.

Solution
==============

One shall use a mechanism similar to the one used in stop, which is to
wait until an acceptable point. It was chosen to be the point in which
the application receives the view and just before starting Recovery.

This brings side effects to it, since Recovery is ready to a certain
extent to deal with these modifications but, during the initial critical
step of Recovery while it is connecting to a donor, the Cluster
information might change, thus rendering Recovery with a "dirty read".

This is caused by delayed information caused by the STOP directive,
which changes Cluster information before informing Recovery that it
should stop.

Recovery will then extract a snapshot copy of the Cluster Manager member
and revert to the old behaviour of failing later in the process
when connecting to a donor.