Bug#20455184: Assertion failed: join_cond in optimizer.cc
The problem query performs a NATURAL LEFT JOIN between a derived table
and a regular table, before applying a NATURAL LEFT JOIN with a
second regular table. The first natural join has no common columns in
the two tables, so there is no generated join condition. In such case,
store_top_level_join_columns() (line 8541) will add TRUE as join
condition to the inner table of the outer join, since an outer join
should, by convention, contain a non-empty join condition.

But the criterion for adding a join condition is:

    if (table_ref_2->outer_join &&
        !table_ref_1->join_cond() && !table_ref_2->join_cond())
      table_ref_2->set_join_cond(new Item_int((longlong) 1,1)); // Always true.

As we see, it also checks whether the join condition of the outer table
is non-empty, which is not the case here, since a WHERE clause of
the derived table has been promoted to become the "join condition"
of this table. Hence, the TRUE condition is not appended to the join,
and subsequent code fails to recognize this as a proper outer join,
which eventually causes an assertion.

The fix is to modify the test to exclude the reference to
table_ref_1->join_cond(), since this is not related to the outer join.