BUG#19910140: HANDLE_FATAL_SIGNAL (SIG=11) IN               REPLACE_USER_TABLE | SQL/SQL_ACL.CC:3125
Description:- Creating a user after altering the user table
definition results in a server crash.

Analysis:- Altering mysql user table definition by dropping
some columns and trying to create a user will result in a
server crash. This is due to the invalid memory access for
the dropped columns of the user table while accesing the
user table.

Fix:- In "replace_user_table()", a condition is added which
will check whether the total number of fields is greater
than or equal to MYSQL_USER_FIELD_AUTHENTICATION_STRING
before storing plugin and authentiacation string values. If
total number of fields is less than 42, an error is thrown,
"ERROR 1054 (42S22): Unknown column 'plugin' in 'mysql.user'".
Similar check is added in "update_user_table()" also.

NOTE:- This patch is only for mysql-5.6. In mysql-5.7 and
mysql-trunk, the bug does not exist anymore.