WL#8313: Set ROW based binary log format by default
Make the following compiled-in default change:
binlog-format=ROW

All test cases outside binlog or rpl test suites that required binlog
and statement based binlog format (previous default format) were changed
to force the required binlog_format, or else these test cases would never
run.

It was found and fixed an issue in sql_delete.cc were it was avoiding the
use of the optimized "delete_all_rows()" for binlog_format = ROW even when
the binary log was disabled or the user disabled the binlog in the session.