Fixed Bug#19695490: CRASH IN CREATE_REF_FOR_KEY ON SELECT + JOIN + UTF8 COLUMN                       + DATETIME INDEX
After materialized lookup is used in semijoin, the server needs build 'ref'
access on materialized temporary table. If temporary table uses hash_field
as the index, the server still uses the materialized columns to look for
index will be impossible. The solution for this problem is to disable
materialized lookup when hash_field is needed.