Bug#18165029: MEMORY LEAK IN MY_GETOPT.CC CAUSE FAILURE OF MYSQLBINLOG WITH REWRITE-DB OPTION
Problem:
========
memory leak on execution of mysqlbinlog with
rewrite-db option.

Analysis:
========
When rewrite-db option is specified in mysqlbinlog tool,
memory is dynamically allocated for two global variables
named rewrite and binlog_rewrite_db. This dynamically
allocated memory is not freed as part of cleanup which
leads to memory leak.

Fix:
===
Added code to cleanup the dynamically allocated memory.