Bug#20458574 FIX ALL THE ARRAY BOUNDS WARNINGS
Patch for 5.7
Remove compiler warnings when building in optimized mode:

In strings/decimal.c fix numerous warnings of the type:
array subscript is below array bounds [-Werror=array-bounds]
This part of the patch is based on a patch
from https://github.com/webscalesql/webscalesql-5.6
That patch added some bounds-checks for loops, and worked for
gcc4.8.3 but not for gcc 4.9.1 so it was re-written to unroll the loops instead.

In other source files, fix numerous warnings of the type:
variable XX set but not used [-Werror=unused-but-set-variable]
variable XX may be used uninitialized in this function [-Werror=maybe-uninitialized]
These were variables used only for debugging purposes,
or variables not proven to be set by all execution paths.

Compiled with gcc 4.7.2 4.8.3 4.9.1 in optimized mode, with -Wall -Werror