BUG#18360716: ERRORS IN MTS WORKER THREADS REPORTED WRONGLY FOR ANONYMOUS TRANSACTIONS
Before this patch, error messages from MTS would use an empty string
when referring to the GTID of an anonymous transaction. The correct
way to refer to anonymous transactions is to use the text "ANONYMOUS".
Empty string is generally used when referring to no GTID at all
(e.g., when GTID_EXECUTED is empty); a single transaction that does
not have a GTID is different from no transaction.

In order to prevent similar bugs in the future, we also simplify the
locking code for Gtid_specification::to_string so that it is very easy
to use the existing function that generates the string for "ANONYMOUS".

This is already tested by rpl_perfschema_execute_status_by_worker, so
we can just update the result file and don't need to write a new test.

@sql/rpl_rli_pdb.cc
- Use the new, simplified version of
  Gtid_specification::to_string.  This ensures that we
  use "ANONYMOUS" for anonymous transactions, rather than
  just "".

@sql/rpl_gtid.h
- Improve some (related) comments.

@mysql-test/suite/rpl/r/rpl_perfschema_applier_status_by_worker.result
- Update result file.

@mysql-test/suite/rpl/t/rpl_perfschema_applier_status_by_worker.test
- Add bug reference.