Bug #20527363	TRUNCATE TEMPORARY TABLE CRASH: 		!DICT_TF2_FLAG_IS_SET(TABLE, DICT_TF2_TEMPORARY)
Problem:

   Server asserts while truncating the temporary tables.

Reason:

If temporary table object is evicted from data dictionary cache during its
lifetime on reload from SYS_TABLES object is not full constructed as it was
when it was created. This is because not all fields of temporary tables are
persisted to SYS_TABLES and dir_path_to_temp_table is one such field which
on reload is set to NULL. Making any decision based on such fields can result
in inconsistent output. Truncate code was trying to do the same.

Reviewed-by: Krunal Bauskar <krunal.bauskar@oracle.com>
Reviewed-by: Sunny Bains <sunny.bains@oracle.com>
RB: 8081