WL#7725 FIREWALL
An audit plugin which use the P_S digset to record SQL statements
and firewall against SQL injections.

Bug#20567900 PREPARED STATEMENTS WON'T ALLOW FOR AUDIT PLUGINS TO RETURN AN ERROR

During the prepare phase of a prepared statement any audit plugin might
raise an error. This error was ignored by the Prepare_statement::prepare
function.

The fix is to check the state of the DA and set the error state
accordingly.