BUG#20190520 - MISPLACED ROW IN TBL PARTITIOND BY KEY(), ALTER TABLE
Problem :
---------
For a table partitioned by Primary Key with default column list[...
partition by key() ...], if the Primary Key is altered [... DROP PRIMARY
KEY, ADD PRIMARY KEY ] table data needs to be redistributed across the
partitions. Currently we do an "inplace alter" for this case and the
table rows are not re-partitioned.

Solution :
----------
The issue is fixed by putting a check to disallow INPLACE for Primary Key
add/Drop if the table is partitioned on Key.

Reviewed-by: Mattias Jonsson <mattias.jonsson@oracle.com>

RB: 8052